Implement the Storage class
Create a Storage class to manage the saving and retrieval of data to and
from the hard disk.

Let's extract the launchDuke, addDBEntry, deleteDBEntry and
setDBEntryDone methods to the Storage class and give Duke a Storage
attribute to allow it to manage storage with the hard disk.
