+++ src/main/java/Duke.java
@@ -1,9 +1,3 @@
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeParseException;
@@ -14,17 +8,19 @@ import java.util.Scanner;
 public class Duke {
     
     private Ui ui;
+    private Storage storage;
     
-    public Duke() {
+    public Duke(String filePath) {
         ui = new Ui();
+        storage = new Storage(filePath, ui);
     }
     
     public static void main(String[] args) {
-        new Duke().run();
+        new Duke("data/duke.txt").run();
     }
     
     public void run() {
-        ArrayList<Task> taskList = launchDuke();
+        ArrayList<Task> taskList = storage.load();
 
         Scanner input = new Scanner(System.in);
         boolean isBye = false;
@@ -55,7 +51,7 @@ public class Duke {
                             throw new DukeException(DukeExceptionType.INVALIDINDEX);
                         } else {
                             Task toSetDone = taskList.get(toSet - 1);
-                            setDBEntryDone(toSetDone.databaseEntry());
+                            storage.setDBEntryDone(toSetDone.databaseEntry());
                             toSetDone.setDone();
                             ui.showDone(toSetDone);
                         }
@@ -70,7 +66,7 @@ public class Duke {
                             throw new DukeException(DukeExceptionType.INVALIDINDEX);
                         } else {
                             Task deleted = taskList.remove(toDelete - 1);
-                            deleteDBEntry(deleted.databaseEntry());
+                            storage.deleteDBEntry(deleted.databaseEntry());
                             ui.showDelete(deleted, --listLength);
                         }
                     }
@@ -156,7 +152,7 @@ public class Duke {
                         }
                         // add task to taskList
                         taskList.add(listLength++, newTask);
-                        addDBEntry(newTask.databaseEntry());
+                        storage.addDBEntry(newTask.databaseEntry());
                         ui.showAdd(newTask, listLength);
                     }
                 }
@@ -176,155 +172,5 @@ public class Duke {
         }
 
         input.close();
-    }
-    
-    public ArrayList<Task> launchDuke() {
-        ArrayList<Task> savedTasks = new ArrayList<>(100);
-        
-        try {
-            File data = new File("data");
-            File duke = new File("data/duke.txt");
-
-            ui.showOpenLine();
-            ui.showInitialise();
-            
-            if (data.mkdir()) {
-                ui.showNewDataDirectory();
-            }
-            if (duke.createNewFile()) {
-                ui.showNewHardDisk();
-            }
-            
-            ui.showCloseLine();
-            
-            BufferedReader reader = new BufferedReader(new FileReader(duke));
-            String currLine = reader.readLine();
-
-            while (currLine != null) {
-                String[] taskString = currLine.split(" \\| ", 3);
-                Task newTask;
-                
-                if (taskString[0].equals("D")) {
-                    String[] deadlineDetails = taskString[2].split(" ");
-                    if (deadlineDetails.length == 2) {
-                        newTask = new Deadline(deadlineDetails[0], LocalDate.parse(deadlineDetails[1]));
-                    } else if (deadlineDetails.length == 3) {
-                        newTask = new Deadline(deadlineDetails[0], 
-                                LocalDate.parse(deadlineDetails[1]), LocalTime.parse(deadlineDetails[2]));
-                    } else {
-                        throw new DukeException(DukeExceptionType.DB_READ);
-                    }
-                    
-                } else if (taskString[0].equals("E")) {
-                    String[] periodDetails = taskString[2].split(" ");
-                    if (periodDetails.length == 3) {
-                        newTask = new Event(periodDetails[0], 
-                                LocalDate.parse(periodDetails[1]), LocalDate.parse(periodDetails[2]));
-                    } else if (periodDetails.length == 4) {
-                        newTask = new Event(periodDetails[0], LocalDate.parse(periodDetails[1]), 
-                                LocalTime.parse(periodDetails[2]), LocalTime.parse(periodDetails[3]));
-                    } else if (periodDetails.length == 5) {
-                        newTask = new Event(periodDetails[0], 
-                                LocalDate.parse(periodDetails[1]), LocalTime.parse(periodDetails[2]), 
-                                LocalDate.parse(periodDetails[3]), LocalTime.parse(periodDetails[4]));
-                    } else {
-                        throw new DukeException(DukeExceptionType.DB_READ);
-                    }
-                    
-                    
-                } else if (taskString[0].equals("T")) {
-                    newTask = new Todo(taskString[2]);
-
-                } else {
-                    throw new DukeException(DukeExceptionType.DB_READ);
-                }
-                
-                if (Integer.parseInt(taskString[1]) == 1) {
-                    newTask.setDone();
-                }
-                
-                savedTasks.add(newTask);
-                currLine = reader.readLine();
-            }
-
-        } catch (DukeException e) {
-            ui.showException(e);
-            
-        } catch (IOException e) {
-            ui.showException(new DukeException(DukeExceptionType.DB_LAUNCH));
-        }
-        
-        return savedTasks;
-    }
-
-    public void addDBEntry(String s) {
-        try {
-            File duke = new File("data/duke.txt");
-            FileWriter writer = new FileWriter(duke, true);
-            writer.write(s + "\n");
-            writer.close();
-            
-        } catch (IOException e){
-            ui.showException(new DukeException(DukeExceptionType.DB_ADD));
-        }
-    }
-
-    public void setDBEntryDone(String s) {
-        try {
-            File updated = new File("data/updated.txt");
-            File duke = new File("data/duke.txt");
-            updated.createNewFile();
-
-            FileWriter writer = new FileWriter(updated, true);
-
-            BufferedReader reader = new BufferedReader(new FileReader(duke));
-            String currLine = reader.readLine();
-
-            while (currLine != null) {
-                if (currLine.equals(s)) {
-                    String[] toSetDone = currLine.split(" \\| ", 3);
-                    writer.write(toSetDone[0] + " | 1 | " + toSetDone[2] + "\n");
-                } else {
-                    writer.write(currLine + "\n");
-                }
-                currLine = reader.readLine();
-            }
-
-            writer.close();
-
-            duke.delete();
-            updated.renameTo(duke);
-            
-        } catch (IOException e) {
-            ui.showException(new DukeException(DukeExceptionType.DB_DONE));
-        }
-    }
-
-    public void deleteDBEntry(String s) {
-        try {
-            File updated = new File("data/updated.txt");
-            File duke = new File("data/duke.txt");
-            updated.createNewFile();
-
-            FileWriter writer = new FileWriter(updated, true);
-
-            BufferedReader reader = new BufferedReader(new FileReader(duke));
-            String currLine = reader.readLine();
-
-            while (currLine != null) {
-                if (!currLine.equals(s)) {
-                    writer.write(currLine + "\n");
-                }
-                currLine = reader.readLine();
-            }
-
-            writer.close();
-
-            duke.delete();
-            updated.renameTo(duke);
-            
-        } catch (IOException e) {
-            ui.showException(new DukeException(DukeExceptionType.DB_DELETE));
-        }
     }
 }
